feat(reminders): notification permissions not granted snackbar
GSoC 2025: Review Reminders

If notification permissions are not granted and the user is trying to edit review reminders, we won't directly block them, but they're likely making a mistake. Hence, we show a persistent Snackbar popup that tells them that notification permissions are not granted. The Snackbar has an "Enable" button which, when pressed, calls Permissions's requestPermissionThroughDialogOrSettings to try and grant the notification permission.